pyproject.toml
[tool.poetry]
name = "audioldm_finetuned"
version = "0.2.0"
description = ""
authors = ["anmartinez18"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ruamel-yaml = "^0.18.5"
wget = "^3.2"
kornia = "^0.7.0"
braceexpand = "^0.1.7"
webdataset = "^0.2.75"
hear21passt = {git = "https://github.com/haoheliu/passt_hear21.git"}
wandb = "^0.16.0"
pytorch-lightning = "^2.1.1"
einops = "^0.7.0"
h5py = "^3.10.0"
transformers = "4.30.2"
ftfy = "^6.1.1"
pandas = "^2.1.3"
matplotlib = "^3.8.1"
taming-transformers-rom1504 = "^0.0.6"
ipdb = "^0.13.13"
langdetect     = "*"
deep-translator = "*"     
mistralai      = "*"
gradio   = "^5.34"
typing-extensions  = "^4.10"
scipy = "^1.11.4"
librosa = "*"
soundfile = "^0.12"
torchlibrosa = { git = "https://github.com/qiuqiangkong/torchlibrosa.git" }


[tool.poetry.scripts]
get-checkpoints = "src.get_checkpoints:main"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
